implementati impostazioni base

+++ include/wikictl.h
@@ -1,14 +1,12 @@
 #ifndef _WIKICTL_H
 #define _WIKICTL_H
 
-#include <stdint.h>
-
 typedef struct {
     char url[100];
     char port[5];
     char endpoint[100];
     char http_request[500];
-    uint8_t show_help;
+    int  show_help;
 } parameters_t;
 
 #endif
